services/core-stack-optimizer/requirements.txt
Flask~=2.3.3
gunicorn==21.2.0
PyYAML==6.0.1
requests==2.31.0
pandas==2.0.3
gevent==23.7.0
SQLAlchemy~=2.0.20
logger~=1.4
loguru~=0.7.1
PyMySQL~=1.1.0
marshmallow-sqlalchemy~=0.29.0
google-cloud-storage
google-cloud-logging
plotly==5.14.1
matplotlib